refactor(engine): decompose buffer_pool/tests.rs by concern
The buffer_pool test file was a 2761-line monolith covering eleven
distinct behavioural surfaces (basic acquire/release, adaptive sizing,
contention, custom allocators, memory cap, throughput tracker, TLS
cache, adaptive pool resizing, telemetry, buffer controller, byte
budget, per-thread slab). Split into focused submodules under
buffer_pool/tests/, all well below the LoC cap, with shared test
allocators promoted to a support module.

No public API change. All 147 tests preserved verbatim.
